+++ R/compare.R
@@ -13,9 +13,14 @@ chunkVector <- function(x, nElems) {
     split(x, factor(sort(rank(x) %% groups)))
 }
 
-processChunk <- function(chunk, data, FUN, ..., progress) {
-    zscores <- loadCMapZscores(data[ , chunk], verbose=FALSE)
-    res <- FUN(zscores, chunk, ..., progress=progress)
+processChunk <- function(chunk, data, FUN, ..., genes=TRUE, progress) {
+    if (is(data, "perturbationChanges")) {
+        loaded <- loadCMapZscores(data[genes, chunk], verbose=FALSE)
+    } else if (is(data, "expressionDrugSensitivityAssociation")) {
+        loaded <- readExpressionDrugSensitivityCorHDF5(
+            data, rows=genes, cols=chunk, loadValues=TRUE)
+    }
+    res <- FUN(loaded, chunk, ..., genes=genes, progress=progress)
     return(res)
 }
 
@@ -32,17 +37,25 @@ processChunk <- function(chunk, data, FUN, ..., progress) {
 #' @param FUN Function: function to run for each chunk
 #' @param num Numeric: numbers of methods to run per chunk
 #' @param ... Arguments passed to \code{FUN}
+#' @param genes Character: genes to compare
 #' @param chunkSize Integer: number of columns to load on-demand (a higher value
 #'   increases RAM usage, but decreases running time)
 #' @inheritParams compareWithAllMethods
 #'
 #' @return Results of running \code{FUN}
 #' @keywords internal
-processByChunks <- function(data, FUN, num, ..., threads=1, chunkSize=10000) {
+processByChunks <- function(data, FUN, num, ..., genes=TRUE, threads=1,
+                            chunkSize=10000) {
     loadFromFile <- is.character(data)
     if (loadFromFile && !file.exists(data)) {
-        msg <- "%s not found: has the CMap z-scores file been moved or deleted?"
-        stop(sprintf(msg, data))
+        if (is(data, "perturbationChanges")) {
+            type <- "z-scores"
+        } else if (is(data, "expressionDrugSensitivityAssociation")) {
+            type <- "gene expression and drug sensitivity association"
+        }
+        source <- attr(data, "source")
+        msg <- "%s not found: has the %s %s file been moved or deleted?"
+        stop(sprintf(msg, data, source, type))
     }
 
     # Display progress per chunk (if multi-threaded and on-demand file loading)
@@ -53,8 +66,8 @@ processByChunks <- function(data, FUN, num, ..., threads=1, chunkSize=10000) {
     if (loadFromFile) {
         chunks <- chunkVector(colnames(data), chunkSize)
         if (threads > 1) pb <- startpb(max=length(chunks))
-        resTmp <- lapply(chunks, processChunk, data, FUN, ..., threads=threads,
-                         progress=pb)
+        resTmp <- lapply(chunks, processChunk, data, FUN, ..., genes=genes,
+                         threads=threads, progress=pb)
         names(resTmp) <- NULL
 
         # Organise lists by the results of each method
@@ -65,7 +78,8 @@ processByChunks <- function(data, FUN, num, ..., threads=1, chunkSize=10000) {
         groups          <- factor(rep(methods, len), unique(methods))
         res             <- split(unlist(unlisted, recursive=FALSE), groups)
     } else {
-        res <- FUN(data, colnames(data), ..., threads=threads, progress=pb)
+        res <- FUN(data, colnames(data), ..., genes=genes, threads=threads,
+                   progress=pb)
     }
     if (!is.null(pb)) closepb(pb)
     return(res)